add suspense to tag page and made server component
